// types/contracts/common.ts
export type Modifiers = {
	abisType: 'normal' | 'modified'
	eth: string;
	ether: string;
	lqty: string;
	lusd: string;
	trove: string;
};

export type PartialModifiers = {
	abisType: 'normal' | 'modified';
} & Partial<Omit<Modifiers, 'abisType'>>;

export interface TypedDeferredTopicFilter<_TCEvent extends TypedContractEvent>
	extends DeferredTopicFilter {}

export interface TypedContractEvent<
	InputTuple extends Array<any> = any,
	OutputTuple extends Array<any> = any,
	OutputObject = any
> {
	(...args: Partial<InputTuple>): TypedDeferredTopicFilter<
		TypedContractEvent<InputTuple, OutputTuple, OutputObject>
	>;
	name: string;
	fragment: EventFragment;
	getFragment(...args: Partial<InputTuple>): EventFragment;
}

type __TypechainAOutputTuple<T> = T extends TypedContractEvent<
	infer _U,
	infer W
>
	? W
	: never;
type __TypechainOutputObject<T> = T extends TypedContractEvent<
	infer _U,
	infer _W,
	infer V
>
	? V
	: never;

export interface TypedEventLog<TCEvent extends TypedContractEvent>
	extends Omit<EventLog, 'args'> {
	args: __TypechainAOutputTuple<TCEvent> & __TypechainOutputObject<TCEvent>;
}